Gelato: ForwardRequest and TaskStatus calls, and EIP-712 signing for ForwardRequests (#511)
Goal of this PR is to add support for posting ForwardRequest messages to Gelato and obtaining the status of those requests post-submission-to-Gelato.

The key abstractions introduced in this PR are:

ForwardRequestCall (in fwd_req_call.rs): Logic for making a single, individual RPC to the Gelato endpoint responsible for accepting the initial forward request submission, which it then enqueues and handles asynchronously.

fwd_req_sig.rs, which contains an impl of the ethers::types::transaction::eip712::Eip712 trait for ForwardRequestOpArgs. This allows an ethers::signers::Signer to sign our message before we submit it to the relevant Gelato endpoint.

TaskStatusCall in task_status_call.rs: Analogous to ForwardRequestCall in that it encapsulates the logic and state required to make an individual call to a Gelato endpoint to get the status of a previously-submitted request.

Chain in chains.rs: Represents chains that are relevant for Abacus and/or supported by Gelato. Handles conversions between the chain_id integer representation expected by Gelato and other types. Also provides the per-chain addresses we need when signing gelato requests for supported chains.