Update submodules (#1966)
### Description

In `fuel-contracts`, we use the `hyperlane-core` rust package and point
to it via git. When cargo tries to install it, it'll look at the entire
`hyperlane-monorepo` and install any git submodules. Apparently it
breaks if the submodule points to a commit that doesn't belong to a
branch in the repo (e.g. look at the warning of
https://github.com/hyperlane-xyz/hyperlane-token/commit/f40093e or
https://github.com/hyperlane-xyz/hyperlane-app-template/commit/4d7159fe92dbfe814b6de1668f2d91c9d2a2b175).
We should really be using commits from main anyways

### Drive-by changes

n/a

### Related issues

n/a

### Backward compatibility

_Are these changes backward compatible?_

Yes

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_

unit tests