Scraper Refresh (#1932)
### Description

Since we need to re-scrape everything anyway, wanted to go ahead and
make some improvements and bug fixes now.

* Adds code to scrape gas payments from the IGP during the sync loop
* Updates gas payments table to work with new gas payment format
* Fixed support with scraper for new configuration format (adds
configurable list of chains to scrape)
* Adds a new view for messages to display all relevant information
* Adds a new view for gas payments to show total amounts paid for a
given message
* Fixed a bug introduced by seaorm (0.10) where `last_inserted_id`
switched from being the first insert id to being the last
* Pulled in a fix from @serejke in #1925
* Updated seaorm to 0.11.1
* Adds support for sepolia
* Improved logging output at various levels

### Drive-by changes

* Updated fallback provider to still retry even if it only has one url
* Adds support for postgres `numeric` types to store Wei values
(supported as of seaorm 0.11)
* Switches to `bytea` representations of binary values
* Switched from the `chrono` library to `time` (`chrono` is outdated and
uses an unpatched version of `time`)
* Removed redundant timestamp from `message` table
* Cleans up some unused code

### Related issues

- Fixes #1925 
- Fixes #1156

### Backward compatibility

_Are these changes backward compatible?_

No - this has a new database format and for this reason defines entirely
new migrations.

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

Yes - I changed it to `scraper3` so that we can run it along side the
old scraper. It now also lines up with the name in the database. There
is a new env var to specify what chains are scraped
`HYP_BASE_CHAINSTOSCRAPE`.


### Testing

_What kind of testing have these changes undergone?_

Manual - Been running the scraper locally and also deployed a current
version of it to testnet3 for a subset of chains.