Fetch with timeout in Liquidity Layer app (#2071)
### Description

Noticed that the portal relayer had hung on some async task, causing
portal messages to stop being relayed. I'm not confident that fetches
were to blame, but based off the logs it seems most likely that one of
the fetches went wrong and an error was never thrown. So this adds a 10
second timeout for each fetch

### Drive-by changes

n/a

### Related issues

n/a

### Backward compatibility

_Are these changes backward compatible?_

Yes

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_
Ran locally, everything still worked