Minor fixes to get LL deploys running smoothly (#2125)
### Description

This PR makes some minor modifications to the deployment structure for
the Liquidity Layer. The most important thing is the embedding of the
`proxyAdmin` addresses in the LL addresses file which is not pretty, but
I did not want to put it in the SDK right now either, but could be
convinced.

In the process of running this, the LL router IGPs have been updated
which is non-material, but otherwise the deploy script runs as a no-op
now.