Reduce log noise (#2336)
### Description

Seeing a ton of errors for process estimate costs even though they are
fully being handled elsewhere and are also not really an error in the
first place. This makes it harder as an operator to spot real errors so
I think we should eliminate this error log. (We also already log this
error elsewhere so it does not add any new information).


![image](https://github.com/hyperlane-xyz/hyperlane-monorepo/assets/4187749/f149a870-52f3-444c-aa62-84b6aa822f44)

This also targets the `ret` logs produced by tracing since many of them
are unhelpful and most of the cases where we actually need a log we
already have one.

### Drive-by changes

None

### Related issues

### Backward compatibility

_Are these changes backward compatible?_

Yes

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_

None