fix(validator): announce directly if no gas estimate (#2454)
### Description

It has been reported that `contract_call.tx.max_cost()` can fail / not
exist on some networks. When that's the case, default to assuming no
additional validator gas is required and just send the tx. If the tx
fails, report the failure and suggest that gas may be insufficient in
the logged error.

### Drive-by changes

None

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/2426

### Backward compatibility

_Are these changes backward compatible?_

Yes

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None


### Testing

_What kind of testing have these changes undergone?_

None

---------

Co-authored-by: Nam Chu Hoai <nambrot@googlemail.com>