Validators on RC (#2461)
### Description

Allows for deploying validators in the RC context, on both `testnet3`
and `mainnet2`

### Drive-by changes

Initially took the approach in
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/2372 to
refactor out the validator keys in a dedicated file so it's easier to
modify / add in the future. However, these are not expected to ever
change so now the keys are hardcoded in `validators.ts`. There is a
cyclic dependency here, because `scripts/get-key-addresses.ts` depends
on the hardcoded addresses in `validators.ts` to know how many validator
addresses to read from AWS, but to know which addresses to hardcode you
need to first run the `get-key-addresses` script. This PR doesn't fix it
but I'm mentioning it for future reference.

Also copy-pastes a change from one of @tkporter's open PRs in
`typescript/infra/src/agents/aws/validator-user.ts`, to get around an
aws bucket creation error.

### Related issues

- Fixes https://github.com/hyperlane-xyz/issues/issues/485

### Backward compatibility

_Are these changes backward compatible?_

Yes

_Are there any infrastructure implications, e.g. changes that would
prohibit deploying older commits using this infra tooling?_

None that I know of - but this PR does impact deployments in general.
Maybe double check there's enough memory and other hardware resources to
support the new validators?


### Testing

_What kind of testing have these changes undergone?_

Manual - checked the kubernetes pods and S3 checkpoints.

---------

Co-authored-by: Trevor Porter <trkporter@ucdavis.edu>
Co-authored-by: Yorke Rhodes <yorke@hyperlane.xyz>
Co-authored-by: Asa Oines <asaoines@gmail.com>