Fix hardhat test script in typescript/sdk/package.json (#2483)
### Description

The previous `test:hardhat` script was not correctly finding and testing
the `.hardhat-test.ts` files in the `src` directory. This commit updates
the script to correctly find these files and pass them to the `hardhat
test` command. Now, all Hardhat tests in the `src` directory and its
subdirectories will be correctly run when the `test:hardhat` script is
used.

### Testing

Unit Tests

Co-authored-by: Yorke Rhodes <yorke@hyperlane.xyz>