Rust updates (#2619)
### Description

Let this one slip a while. Updates our version of rust and as many
packages as did not conflict with solana. I think we are going to need
to figure something better out going forward with solana since if we
have a couple more envs like this, it is going to make updates very
hard.

### Drive-by changes

None

### Related issues

General Maintenance

### Backward compatibility

Yes

### Testing

Manual & Unit Tests