Add SetInterchainGasPaymaster instruction to warp route programs (#2657)
### Description

Sadly I missed adding an instruction to allow a warp route owner to set
the IGP. This adds that. We're deploying a warp route without an IGP for
now because the UI hasn't been updated to work with IGP payments yet
(which is required as the UI needs to handcraft specific accounts that
are passed in). We'll probably wanna flip the switch to charging gas at
some point, so want to make sure we're able to do that

The way the tests are set up involves a lot of code duplications - sorry
🙈 - we should revisit at some point as now isn't the time

### Drive-by changes

n/a

### Related issues

Related #2613 

### Backward compatibility

Fully backward compatible

### Testing

Unit / functional tests