Agent Config Parsing Casing Fixes (#2666)
### Description

This is a fix for an issue in the arguments and environment where we are
unable to correctly support attribute names due to the casing. The main
issue cannot be fully fixed practically by this PR because it would
cause some configuration cases to break. For now it adds manual
exceptions for the command line argument case and leaves envs as they
were. There are then two new parsers that can be used with the new
config format.

### Drive-by changes

Also adds this to argument parser since it was trivial to copy/paste it
and it will allow us to harden parsing when we switch over.

### Related issues

- Fixes #2662 
- Fixes #2663 
- Progress on #2215 

### Backward compatibility

Yes

### Testing

Unit Tests