Agent transaction overrides, ensure agent configs are updated in CI (#3563)
### Description

- Adds some basic transaction overrides for Ethereum chains to be used
by the agents. This follows the existing config schema in chain metadata
that we have been using on the TS side for a while. The overrides are as
follows:
- `transactionOverrides.gasPrice` - if specified, non EIP-1559 txs are
used, and they use this gasPrice
- `transactionOverrides.gasLimit` - if specified, this takes precedence
over any gas estimation that's done and is the gas limit that's used for
txs. (This is pretty niche and I don't imagine we'll ever wanna use this
internally, but happens to cover the use case described
[here](https://discord.com/channels/935678348330434570/984123861144600587/1224617234580766741),
which is nice to cover)
- `transactionOverrides.maxFeePerGas` - if specified and if the chain
supports EIP 1559 txs, is used for the `maxFeePerGas`
- `transactionOverrides.maxPriorityFeePerGas` - if specified and if the
chain supports EIP 1559 txs, is used for the `maxPriorityFeePerGas`
- Updates some of the transaction overrides in mainnet3 / testnet4 to be
more reasonable now that we'll be using them to submit transactions
pretty frequently
- Note that atm the transaction overrides are most easily applied to the
agents via the config JSONs, and not via our K8s configuration. The
upshot of this is if we wanna change a transaction override right now,
the easiest way to do it will be to run `update-agent-config.ts` and
then build a new image. Created
https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3565 to track
making this experience better
- Part of this change means that it's more important to keep the
`transactionOverrides` specified in the SDK / infra up to date with the
agent config. So I ended up doing
https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3325 along
the way, which also required doing
https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3311.
- To fix the issue with arbitrum blocks, we now require the index.from
to be specified in the chain metadata if the chain's technical stack is
arbitrum nitro
- A new script, `update-agent-config.ts` is added, and it's ran for
mainnet3 and testnet4 in CI to make sure there's no diff (similar idea
to what we do with prettier)
- Along the way I renamed mainnet3_config.json and testnet4_config.json
to mainnet_config.json and testnet_config.json. This has been a source
of confusion with ppl in the past - in v2 we named them without any
numeric suffixes, so we now just do that again

### Drive-by changes

- There's a drive-by to poll more frequently for pending transactions (2
seconds). The default is 7 seconds in Ethers, and for many chains that
have really quick block times, this may be a way for us to speed up us
learning that a transaction is included
- Don't enforce gas on the inevm routes, looks like there's some
weirdness there that was reported by Nam

### Related issues

- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3562
- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3325
- Fixes https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3311

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->