Indexing Type Wrapper (#3610)
### Description

Depends on
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/3544.

Wraps all types to be indexed in an `Indexed<T>` struct in the relayer.
This allows for using sequece-aware cursors for data that doesn't
contain a sequence but can have it inferred from metadata (as is the
case for the IGP on sealevel).

`Indexed<T>` stores the sequence as an `Option`, so the db can have
additional sequence data stored in it when the sequence is defined -
this is how `process_indexed_gas_payment` is implemented now.

In addition, the IGP payments and Deliveries are configured to be
indexed with a `SequenceAware` cursor on sealevel (see
`cursors/mod.rs`).

### Drive-by changes

Implements the custom `Encode` and `Decode` traits for
`IndexingDecorator`, which can be generalized to `Option<T>` where `T:
Encode + Decode`.

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

E2E