fix: fetch the registry's `main` branch in the monorepo Docker image (#3841)
### Description

Infra services require the presence of a registry. This is expected to
be found at the env var REGISTRY_URI. This PR fetches the latest
registry from `main` and uses that as the registry in the monorepo image
for infra services / scripts to make use of. To prevent caching of this
git clone, we use a build-time arg such that if the arg changes, any
later layers in the image will be ran. See
https://stackoverflow.com/a/73501716

### Drive-by changes

<!--
Are there any minor or drive-by changes also included?
-->

### Related issues

- Partially fixes
https://github.com/hyperlane-xyz/hyperlane-monorepo/issues/3827 - we'll
need a deploy to fully close it out

### Backward compatibility

<!--
Are these changes backward compatible? Are there any infrastructure
implications, e.g. changes that would prohibit deploying older commits
using this infra tooling?

Yes/No
-->

### Testing

<!--
What kind of testing have these changes undergone?

None/Manual/Unit Tests
-->