feat: persist status labels (#4043)
### Description

Builds on top of
https://github.com/hyperlane-xyz/hyperlane-monorepo/pull/4003 to persist
updates to an operation's status to the db, so we no longer default to
`FirstPrepareAttempt` upon restarting the relayer

### Drive-by changes

Moves `hyperlane_db` from `hyperlane_base` to `hyperlane_core` so it can
be added to the `PendingOperation` trait. `PendingMessage` already
stores a reference to the db in `ctx`, so I figured it doesn't break any
assumptions.

### Related issues

<!--
- Fixes #[issue number here]
-->

### Backward compatibility
Yes

### Testing

Unit tests encoding of `PendingOperationStatus`, need to manually tests
e2e