feat: include `id` field in `list_operations` endpoint (#4610)
### Description

The `id` field was missing if the serialized `PendingOperation` was a
`PendingMessage`, which made it hard to look up the message in relayer
logs or the explorer. This PR takes advantage of `id()` being a required
method on `PendingOperation` to derive and display the ID as part of
endpoint logic.

### Testing

Unit Tests